+++ src/components/Chat/ChatHeader/styles.scss
